adding view modules separating features into modules versions gradle
@codeAbode

+++ weather/src/main/java/com/cigna/mobile/weather/api/WeatherApi.kt
@@ -7,5 +7,5 @@ import retrofit2.http.Path
 
 interface WeatherApi{
     @GET("points/{lat},{lng}")
-    suspend fun getWeatherByLatLngAsync(@Path("lat") lat: Int, @Path("lng") lng: Int): Deferred<WeatherContext>
+    suspend fun getWeatherByLatLngAsync(@Path("lat") lat: Double, @Path("lng") lng: Double): Deferred<WeatherContext>
 }
